compute hour difference when rotating the bezel
